refactor: TT-401 Rename the time entries folder
