+++ V2/serverless.yml
@@ -38,7 +38,7 @@ package:
 
 functions:
   get_activities:
-    handler: time_entries/interface.get_activities
+    handler: time_tracker/activities/interface.get_activities
     events:
       - http: true
         x-azure-settings:
@@ -48,7 +48,7 @@ functions:
           authLevel: anonymous
 
   delete_activity:
-    handler: time_entries/interface.delete_activity
+    handler: time_tracker/activities/interface.delete_activity
     events:
       - http: true
         x-azure-settings:
@@ -58,7 +58,7 @@ functions:
           authLevel: anonymous
 
   update_activity:
-    handler: time_entries/interface.update_activity
+    handler: time_tracker/activities/interface.update_activity
     events:
       - http: true
         x-azure-settings:
@@ -68,7 +68,7 @@ functions:
           authLevel: anonymous    
 
   create_activity:
-    handler: time_entries/interface.create_activity
+    handler: time_tracker/activities/interface.create_activity
     events:
       - http: true
         x-azure-settings:
