+++ V2/tests/integration/daos/activities_json_dao_test.py
@@ -1,5 +1,5 @@
-from time_entries._infrastructure import ActivitiesJsonDao
-from time_entries._domain import Activity
+from time_tracker.activities._infrastructure import ActivitiesJsonDao
+from time_tracker.activities._domain import Activity
 from faker import Faker
 import json
 import pytest
