+++ V2/tests/unit/use_cases/activities_use_case_test.py
@@ -1,4 +1,4 @@
-from time_entries._domain import _use_cases
+from time_tracker.activities._domain import _use_cases
 from pytest_mock import MockFixture
 from faker import Faker
