+++ V2/time_tracker/activities/_domain/_services/_activity.py
@@ -1,4 +1,4 @@
-from time_entries._domain import ActivitiesDao, Activity
+from time_tracker.activities._domain import ActivitiesDao, Activity
 import typing
