+++ V2/time_tracker/activities/_domain/_use_cases/_update_activity_use_case.py
@@ -1,4 +1,4 @@
-from time_entries._domain import ActivityService, Activity
+from time_tracker.activities._domain import ActivityService, Activity
 
 
 class UpdateActivityUseCase:
