+++ V2/time_tracker/activities/_infrastructure/_data_persistence/_activities_json_dao.py
@@ -1,4 +1,4 @@
-from time_entries._domain import ActivitiesDao, Activity
+from time_tracker.activities._domain import ActivitiesDao, Activity
 import dataclasses
 import json
 import typing
